[LibOS] Move shim's ldsodefs.h to include/elf/ldsodefs.h
As a preparatory step for the next patch, move the shim's ldsodefs.h
to include/elf/ldsodefs.h.
