+++ LibOS/shim/src/elf/shim_rtld.c
@@ -27,7 +27,7 @@
 #include <errno.h>
 
 #include "elf.h"
-#include "ldsodefs.h"
+#include "elf/ldsodefs.h"
 #include "shim_checkpoint.h"
 #include "shim_flags_conv.h"
 #include "shim_fs.h"
